requirements.txt
requests==2.31.0
pandas==2.1.4
pytz==2023.3
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2
sqlalchemy==2.0.23
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.28.2
